fix some merge conflicts and address review comments
