+++ fdbclient/include/fdbclient/Metacluster.h
@@ -28,6 +28,33 @@
 #include "fdbclient/KeyBackedTypes.h"
 #include "flow/flat_buffers.h"
 
+namespace TenantAPI {
+// Represents the various states that a tenant could be in.
+// In a standalone cluster, a tenant should only ever be in the READY state.
+// In a metacluster, a tenant on the management cluster could be in the other states while changes are applied to the
+// data cluster.
+//
+// REGISTERING - the tenant has been created on the management cluster and is being created on the data cluster
+// READY - the tenant has been created on both clusters, is active, and is consistent between the two clusters
+// REMOVING - the tenant has been marked for removal and is being removed on the data cluster
+// UPDATING_CONFIGURATION - the tenant configuration has changed on the management cluster and is being applied to the
+//                          data cluster
+// RENAMING - the tenant is in the process of being renamed
+// ERROR - the tenant is in an error state
+//
+// A tenant in any configuration is allowed to be removed. Only tenants in the READY or UPDATING_CONFIGURATION phases
+// can have their configuration updated. A tenant must not exist or be in the REGISTERING phase to be created. To be
+// renamed, a tenant must be in the READY or RENAMING state. In the latter case, the rename destination must match
+// the original rename attempt.
+//
+// If an operation fails and the tenant is left in a non-ready state, re-running the same operation is legal. If
+// successful, the tenant will return to the READY state.
+enum class TenantState { REGISTERING, READY, REMOVING, UPDATING_CONFIGURATION, RENAMING, ERROR };
+
+std::string tenantStateToString(TenantState tenantState);
+TenantState stringToTenantState(std::string stateStr);
+} // namespace TenantAPI
+
 struct ClusterUsage {
 	int numTenantGroups = 0;
 
@@ -136,6 +163,7 @@ struct MetaclusterTenantMapEntry {
 	}
 
 	bool operator==(MetaclusterTenantMapEntry const& other) const;
+	bool operator!=(MetaclusterTenantMapEntry const& other) const;
 
 	template <class Ar>
 	void serialize(Ar& ar) {
